LocalCommon and Common was separated by historical reason. it's time to be one
