+++ app/models/fluentd/agent/td_agent.rb
@@ -2,7 +2,6 @@ class Fluentd
   class Agent
     class TdAgent
       include Common
-      include LocalCommon
 
       def self.default_options
         {
